Admission Notification 2017
FYJC Pune 2016 Admission Dates will be announced soon…..
The process of filling the online application form began on 1 June 2015 and will end at 5 p.m. on 18 June 2015. The total number of seats in the junior colleges in Pune is 72,105. The last year there were 67,665 seats in the colleges. The first merit list for the admissions will be declared on 24 June at 5 p.m. After the completion of the admission process, the classes for the academic session 2015-16 will begin from 15 July 2015.